Close file pointers used in KLEE automatically/ Add propper failure handling
Close file pointers automatically in case in case of erroneus
situations.

Handle the case a file could not be opened more gracefully (instead of
throwing an assertions).

For test case generation, make sure a file could be opened correctly
instead of just writing into it. This solves crashes of KLEE when it
runs out of file handlers.