Make StatsTracker public and managed
StatsTracker is used beyond just tracking/printing states. It is
essential for different features to work (e.g. searchers).

This patch transform the StatsTracker into a managed one such that it
can be assured, a valid instance is available through KLEE's lifetime.

StatsTracker needs to be initialized after the module has been
initialised.